requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
matplotlib
tqdm
scikit-learn
lxml
ultralytics
imutils
supervision